requirements.txt
together
python-dotenv
pymongo
coloroma
Flask
Flask-Cors